split and refactor setup to engine
