+++ ts/speech_generator/summary_speech_generator.ts
@@ -22,7 +22,7 @@
 
 import { AbstractSpeechGenerator } from './abstract_speech_generator.js';
 import * as SpeechGeneratorUtil from './speech_generator_util.js';
-import { setup as EngineSetup } from '../common/engine_setup.js';
+import { setupEngine } from '../common/engine_setup.js';
 import { Attribute } from '../enrich_mathml/enrich_attr.js';
 
 export class SummarySpeechGenerator extends AbstractSpeechGenerator {
@@ -31,7 +31,7 @@ export class SummarySpeechGenerator extends AbstractSpeechGenerator {
    */
   public getSpeech(node: Element, _xml: Element) {
     // SpeechGeneratorUtil.connectAllMactions(xml, this.getRebuilt().xml);
-    EngineSetup(this.getOptions());
+    setupEngine(this.getOptions());
     const id = node.getAttribute(Attribute.ID);
     const snode = this.getRebuilt().streeRoot.querySelectorAll(
       (x) => x.id.toString() === id
